Split PageSettings/DeckGroup.py into DeckGroupParts/
